cloud_class/user_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
databases
cryptography
prometheus-fastapi-instrumentator
python-logging-loki
